fix Assertion error DeepEP/csrc/kernels/intranode.cu:928: 'false and "Unsupported type"'

+++ vllm/model_executor/layers/fused_moe/deepep_ht_prepare_finalize.py
@@ -337,7 +337,8 @@ class DeepEPHTPrepareAndFinalize(mk.FusedMoEPrepareAndFinalize):
             )
         dbo_yield_and_switch_from_compute_to_comm()
         combined_x, _, event = self.buffer.combine(
-            x=fused_expert_output,
+            # HT combine only supports BF16
+            x=fused_expert_output.to(torch.bfloat16),
             handle=handle,
             topk_weights=None,
             config=self._get_combine_config(),
@@ -355,7 +356,7 @@ class DeepEPHTPrepareAndFinalize(mk.FusedMoEPrepareAndFinalize):
                     event.current_stream_wait()
                 dbo_switch_to_comm()
                 # Respect inplace outputs.
-                output.copy_(combined_x, non_blocking=True)
+                output.copy_(combined_x.to(dtype=output.dtype), non_blocking=True)
 
                 # TODO(lucas): refactor the modular kernel so this will be
                 # handled there
@@ -366,7 +367,7 @@ class DeepEPHTPrepareAndFinalize(mk.FusedMoEPrepareAndFinalize):
             # TODO(lucas): support this case with the refactored modular kernel
             assert not dbo_enabled()
             # Respect inplace outputs.
-            output.copy_(combined_x, non_blocking=True)
+            output.copy_(combined_x.to(dtype=output.dtype), non_blocking=True)
             return None
 
     def finalize_async(
